% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyZillow_main.R
\docType{package}
\name{tidyZillow-package}
\title{Tidy the zillow housing research data}
\description{
tidyZillow quickly tidies the zillow housing research time series
using dplyr and tidyr
}
\details{
The only function you're likely to need from \pkg{tidyZillow} is
\code{\link{tidyZillow}}.
}

