% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyZillow.R
\name{tidyZillow}
\alias{tidyZillow}
\title{Tidy zillow research time series data}
\usage{
tidyZillow(zillow.df, series.name = "value")
}
\arguments{
\item{zillow.df}{A dataframe from with the raw zillow
data}

\item{series.name}{a string with the series name. default is
"value"}
}
\value{
a tidy dataframe with the zillow data
}
\description{
Tidy zillow time series research data
}
\details{
see \url{http://www.zillow.com/research/data/}
}
\examples{
##state level data
library(readr)
zillow.state <- read_csv("http://files.zillowstatic.com/research/public/State/State_Zhvi_AllHomes.csv")
zillow.state <- tidyZillow(zillow.state, series.name = "ZillowState")
head(zillow.state)
}
\author{
Chandler Lutz
}

